03_Code_Templates/agent_project_template/requirements.txt
# Agent 项目依赖模板

# 配置文件处理
pyyaml>=6.0

# 日志
colorlog>=6.7.0  # 彩色日志输出

# 命令行参数
click>=8.1.0  # 更强大的命令行界面（可选）

# 数据处理
pandas>=2.0.0  # 数据处理（可选）
requests>=2.31.0  # HTTP 请求（可选）

# 开发工具
pytest>=7.4.0  # 测试框架
pytest-cov>=4.1.0  # 测试覆盖率
black>=23.7.0  # 代码格式化
flake8>=6.1.0  # 代码检查
mypy>=1.5.0  # 类型检查

# 根据你的项目需求添加其他依赖
# 例如：
# openai>=1.0.0  # OpenAI API
# langchain>=0.0.300  # LangChain
# beautifulsoup4>=4.12.0  # HTML 解析
# selenium>=4.15.0  # Web 自动化
